+++ processing/src/main/java/org/apache/druid/segment/StringDimensionIndexer.java
@@ -208,7 +208,7 @@ public class StringDimensionIndexer extends DictionaryEncodedColumnIndexer<int[]
     // string length is being accounted for each time they are referenced, based on dimension handler interface,
     // even though they are stored just once. It may overestimate the size by a bit, but we wanted to leave
     // more buffer to be safe
-    long estimatedSize = keys.length * Integer.BYTES;
+    long estimatedSize = (long) keys.length * Integer.BYTES;
 
     String[] vals = dimLookup.getValues(keys);
     for (String val : vals) {
